Make slab-reassign free the reassigned slab
Description: This patch is a hack that enables us to us to use memcached with
Ingot.  During the rebalancing process, a malloc'd slab can be placed into
another slab class.  Within the context of Ingot, this will not work - we
allocate slabs to be in a particular slab class, and the var_size metadata is
set to work with that.  So, we actually have to free the class.

Actual information:

Drop slab upon reaching end of reassignment process, as opposed to actually
reassigning the slab.

Patch tested on bb0980fbbafd4eb723f76918e7ca364360315c1b, the patch preceding
the Ingot modifications.  Tested using memtier_benchmark, but briefly.