Comment-out code only used for slab pre-allocation
==============================================
Start: patches not guaranteed to be bisectable
==============================================

Comment-out code that's only used for pre-allocating slabs.  Commenting out the
call to the function whilst leaving the function defined causes the compiler to
complain due to certain flags.

Untested, but I think this is fine, as this patch simply disables a feature.