Changing / fixing various pmalloc/pfree calls
=============================================================================
End: patches not guaranteed to be bisectable

Warning: the work done in this patch is cummulative work, and was most likely
done before the work done in the previous few patches.  As being mentioned
repeatedly, this patch was tested as a whole with the surrounding patches.
=============================================================================

Replace persistent_alloc() with persistent_var_size_alloc(). Changing
memcached's memory_allocate function to include slab class size, to pass this
to Ingot.

Adding the necessary | PT_ARRAY flag to mark things as being an array.

Replacing call to free with a call to persistent_free() (should have done this
previously).